pyproject.toml
[project]
name = "content-project"
version = "0.1.0"
description = ""
authors = [
    {name = "NikolaiIakovlev",email = "118279320+NikolaiIakovlev@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "django (>=4.2,<5.0)",
    "djangorestframework (>=3.15,<4.0)",
    "psycopg (>=3.1,<4.0)",
    "celery (>=5.3,<6.0)",
    "redis (>=4.6,<5.0)",
    "flower (>=1.2,<2.0)",
    "django-admin-sortable2 (>=2.2.8,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"
pytest = "^8.4.1"
pytest-django = "^4.11.1"

